+++ R/get_solution.R
@@ -15,6 +15,11 @@
 #'   (and some preprocessed new columns).
 #' @examples
 #' \donttest{
+#' \dontshow{
+#' # set the cachedir to a temporary directory
+#' pth <- withr::local_tempdir(pattern = "snvecR")
+#' withr::local_options(snvecR.cachedir = pth)
+#' }
 #' get_solution("full-ZB18a")
 #' get_solution("ZB20a")
 #' get_solution("La11")
@@ -29,6 +34,8 @@ get_solution <- function(astronomical_solution = "full-ZB18a", quiet = FALSE, fo
   solutions <- c("full-ZB18a", # the default
                  # special case to catch ambiguous naming
                  "ZB18a",
+                 "full-La10",
+                 "full-La10a", "full-La10b", "full-La10c", "full-La10d",
                  "full-La11", # just to thow an error
                  # we rely on astrochron to get these for us
                  "La04", "La10a", "La10b", "La10c", "La10d", "La11",
@@ -41,36 +48,49 @@ get_solution <- function(astronomical_solution = "full-ZB18a", quiet = FALSE, fo
                  "ZB20a", "ZB20b", "ZB20c", "ZB20d")
 
   if (!astronomical_solution %in% solutions) {
-    cli::cli_abort(c("{.var astronomical_solution} must be one of: {.or {.q {solutions}}}",
-                     "x" = "You've supplied {.q {astronomical_solution}}"))
+    cli::cli_abort(c(
+      "{.var astronomical_solution} must be one of: {.or {.q {solutions}}}",
+      "x" = "You've supplied {.q {astronomical_solution}}"
+    ))
   }
 
   if (astronomical_solution == "ZB18a") {
-    cli::cli_abort(c("i" = "There are multiple versions of {.q ZB18a} available.",
-                     "!" = "Please select the one you want:",
-                     "*" = "{.q full-ZB18a}: for the precession-tilt solutions",
-                     "*" = "{.q ZB18a-100}: for the eccentricity and inclination for the past 100 Ma",
-"*" = "{.q ZB18a-300}: for the eccentricity and inclination for the past 300 Ma"))
+    cli::cli_abort(c(
+      "There are multiple versions of {.q ZB18a} available.",
+      "!" = "Please select the one you want:",
+      "*" = "{.q full-ZB18a}: for the full astronomical solution.",
+      "*" = "{.q ZB18a-100}: for the eccentricity and inclination for the past 100 Myr.",
+      "*" = "{.q ZB18a-300}: for the eccentricity and inclination for the past 300 Myr."
+    ))
   }
 
   if (grepl("^full-La", astronomical_solution)) {
-    cli::cli_abort(c("i" = "Astronomical solution {.q {astronomical_solution}} is not supported.",
-                     "!" = "The input OS for snvec must be either in the:",
-                     "*" = "Heliocentric inertial reference frame (HCI)",
-                     "*" = "Ecliptic reference frame (J2000).",
-                     "x" = "The La10 and La11 solutions are in the invariant/inertial reference frame.",
-                     "i" = "To resolve this, you need the positions/velocities and masses of all the bodies.",
-                     "i" = "Or the angles between their inertial reference frame and J2000.",
-                     "i" = "Pull requests welcome."))
+    cli::cli_abort(c(
+      "Astronomical solution {.q {astronomical_solution}} is not supported.",
+      "!" = "The input astronomical solution for {.fun snvec} must be either in the:",
+      "*" = "Heliocentric inertial reference frame (HCI)",
+      "*" = "Ecliptic reference frame (J2000).",
+      "x" = "The La10x and La11 solutions are in the invariant/inertial reference frame.",
+      "i" = "To resolve this, you need the positions/velocities and masses of all the bodies.",
+      "i" = "Or the angles between their inertial reference frame and J2000.",
+      "i" = "Pull requests welcome."
+    ))
   }
 
-  if (grepl("^La[0-9][a-z]?", astronomical_solution)) {
-    cli::cli_warn(c("i" = "Relying on {.pkg astrochron} to get solution {.q {astronomical_solution}}",
+  if (grepl("^La[0-9][0-9][a-z]?", astronomical_solution)) {
+    cli::cli_warn(c(
+      "i" = "Relying on {.pkg astrochron} to get solution {.q {astronomical_solution}}",
                     "i" = "We do not cache these results.",
                     "!" = "{.pkg astrochron} converts time from -kyr to ka by default."))
-    rlang::check_installed("astrochron", reason = "to use `astrochron::getLaskar()`")
-    dat <- tibble::as_tibble(astrochron::getLaskar(sol = tolower(astronomical_solution)))
-    cli::cli_warn(c("i" = "Output has column names {.q {colnames(dat)}}"))
+    rlang::check_installed("astrochron",
+                           reason = "to use `astrochron::getLaskar()`")
+    dat <- tibble::as_tibble(
+      astrochron::getLaskar(sol = tolower(astronomical_solution),
+                            verbose = !quiet)
+    )
+    cli::cli_warn(c(
+      "i" = "Output has column names {.q {colnames(dat)}}"
+    ))
   }
 
   if (astronomical_solution == "full-ZB18a" ||
@@ -107,8 +127,8 @@ get_ZB <- function(astronomical_solution = "full-ZB18a",
                    quiet = FALSE,
                    force = FALSE) {
   base_url <- "http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/"
-  cachedir <- tools::R_user_dir("snvecR", which = "cache")
-
+  # get cachedir from globals.R
+  ## cachedir <- R_user_dir("snvecR", which = "cache")
   if (astronomical_solution == "full-ZB18a") {
     url <- paste0(base_url, "PrecTilt/OS/ZB18a/ems-plan3.dat")
     raw_col_names <- c("t", # time in days
@@ -134,9 +154,9 @@ get_ZB <- function(astronomical_solution = "full-ZB18a",
   }
 
   # where will we save our cached results?
-  raw_path <- paste0(cachedir, "/", astronomical_solution, ".dat")
-  csv_path <- paste0(cachedir, "/", astronomical_solution, ".csv")
-  rds_path <- paste0(cachedir, "/", astronomical_solution, ".rds")
+  raw_path <- paste0(cachedir(), "/", astronomical_solution, ".dat")
+  csv_path <- paste0(cachedir(), "/", astronomical_solution, ".csv")
+  rds_path <- paste0(cachedir(), "/", astronomical_solution, ".rds")
 
   # read final processed file from cache if available
   if (!force && file.exists(rds_path)) {
@@ -167,19 +187,26 @@ get_ZB <- function(astronomical_solution = "full-ZB18a",
       }
     }
   } else {# files don't exist or force
-    if (!quiet) cli::cli_alert_info("The astronomical solution {astronomical_solution} has not been downloaded.")
-
+    if (!quiet) {
+      cli::cli_alert_info(c(
+        "!" = "The astronomical solution {.q {astronomical_solution}} has not been downloaded."
+      ))
+    }
     # default to downloading/caching if not interactive (i.e. GitHub actions)
     if (force || !interactive()) {
       download <- TRUE
       save_cache <- TRUE
     } else {
       # a logical, TRUE if Yes, no if otherwise
-      download <- utils::menu(c("Yes", "No"),
-                              title = "Would you like to download and process it now?") == 1L
+      download <- utils::askYesNo("Would you like to download and process it now?")
+      if (is.na(download)) {
+        cli::cli_abort("Cancelled by user.", call = NULL)
+      }
       if (download) {
-        save_cache <- utils::menu(c("Yes", "No"),
-                                  title = "Would you like to save the results to .csv and .rds?") == 1L
+        save_cache <- utils::askYesNo("Would you like to save the astronomical solution to .csv and .rds?")
+        if (is.na(save_cache)) {
+          cli::cli_abort("Cancelled by user.", call = NULL)
+        }
       } else {
         save_cache <- FALSE
       }
@@ -191,7 +218,11 @@ get_ZB <- function(astronomical_solution = "full-ZB18a",
     }
 
     # read the file from the website
-    if (!quiet) cli::cli_alert_info("Reading {.file {basename(raw_path)}} from website {.url {url}}.")
+    if (!quiet) {
+      cli::cli_alert_info(
+        "Reading {.file {basename(raw_path)}} from website {.url {url}}."
+      )
+    }
     if (astronomical_solution == "full-ZB18a") {
       raw <- readr::read_table(url,
                                col_names = raw_col_names,
@@ -212,32 +243,49 @@ get_ZB <- function(astronomical_solution = "full-ZB18a",
     if (!save_cache) {
       return(raw)
     } else {
-      if (!dir.exists(cachedir)) {
-        dir.create(cachedir, recursive = TRUE, showWarnings = TRUE)
+      if (!dir.exists(cachedir())) {
+        dir.create(cachedir(), recursive = TRUE, showWarnings = TRUE)
       }
-      if (!quiet) cli::cli_alert_info("The cache directory is {.file {cachedir}}.")
+      if (!quiet) {
+        cli::cli_alert_info(
+          "The cache directory is {.file {cachedir()}}."
+        )
+      }
 
       # also copy the raw file to disk
       # even though we've read it in using read_table directly
 
       if (rlang::is_installed("curl")) {
         curl::curl_download(url, destfile = raw_path)
-        if (!quiet) cli::cli_alert_info("Saved {.file {basename(raw_path)}} to cache.")
+        if (!quiet) {
+          cli::cli_alert_info(
+            "Saved {.file {basename(raw_path)}} to cache."
+          )
+        }
       } else {
-        if (!quiet) cli::cli_alert_info("i" = "Did not download {basename(raw_path)} to cache",
-                                        "!" = "Install {.pkg curl} and re-run with {force = TRUE} if you want to.")
+        cli::cli_warn(c(
+          "i" = "Did not download the raw data file {.file {
+basename(raw_path)}} to cache.",
+          "!" = "If you want to do so, install the {.pkg curl} package and re-run with `force = TRUE`"
+        ))
       }
-
 
       # write intermediate result to csv
       readr::write_csv(raw, csv_path)
-      if (!quiet) cli::cli_alert_info("Saved cleaned-up {.file {basename(csv_path)}} to cache.")
+      if (!quiet) {
+        cli::cli_alert_info(
+        "Saved cleaned-up {.file {basename(csv_path)}} to cache."
+        )
+      }
 
       # write final result to rds cache
       readr::write_rds(raw, rds_path)
       if (!quiet) {
-        cli::cli_alert("Saved solution with helper columns {.file {basename(rds_path)}} to cache.",
-                       "i" = "Future runs will read from the cache, unless you specify `force = TRUE`.")
+        cli::cli_inform(c(
+          "i" = "Saved astronomical solution with helper columns {.file {basename(rds_path)}} to cache.",
+          "i" = "Future calls to `get_solution({.q {astronomical_solution}})` will read from the cache.",
+          "!" = "If you don't want this, specify `force = TRUE`."
+        ))
       }
     }
     return(raw)